moved components to general
